# openapi/openapi.yaml
openapi: 3.0.0
info:
  title: Workk API
  version: 1.0.0
...
paths:
...
  /matches:
    get:
      summary: Get all matches
      description: Returns list of all matches in the system.
      tags:
        - matches
      responses:
        200:
          description: Successfully fetched matches.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MatchDTO'
        401:
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        500:
          description: Internal Server Error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /matches/{id}:
    get:
      summary: Get match by ID
      description: Retrieves detailed information about a specific match by their unique ID.
      tags:
        - matches
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        200:
          description: Match found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MatchDTO'
        400:
          description: Bad request.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        401:
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        404:
          description: Match not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        500:
          description: Internal Server Error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /likes/like:
    post:
      summary: Like a freelancer or project
      description: Adds a like from one user to another user. Returns the created like and optionally the next suggestion.
      tags:
        - likes
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LikeRequest'
      responses:
        200:
          description: Like successfully created, get next card.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LikeResponse'
        400:
          description: Bad request.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        401:
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        404:
          description: User not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        500:
          description: Internal Server Error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /likes/dislike:
    post:
      summary: Dislike a freelancer or project
      description: Marks a user or project as disliked. Returns the next suggestion.
      tags:
        - likes
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LikeRequest'
      responses:
        200:
          description: Dislike recorded successfully, get next card.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DislikeResponse'
        400:
          description: Bad request.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        401:
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        404:
          description: User not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        500:
          description: Internal Server Error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  schemas:
...
    MatchDTO:
      x-go-type: dto.MatchDTO
      x-go-type-import:
        name: dto
        path: github.com/kitoyanok66/workk/dto
      type: object
      description: Data Transfer Object representing a match between freelancer and project
      properties:
        id:
          type: string
          format: uuid
        freelancer:
          $ref: '#/components/schemas/FreelancerDTO'
        project:
          $ref: '#/components/schemas/ProjectDTO'
        freelancer_user:
          $ref: '#/components/schemas/UserDTO'
        project_user:
          $ref: '#/components/schemas/UserDTO'
        created_at:
          type: string
          format: date-time
    LikeDTO:
      x-go-type: dto.LikeDTO
      x-go-type-import:
        name: dto
        path: github.com/kitoyanok66/workk/dto
      type: object
      description: Data Transfer Object representing a like.
      properties:
        id:
          type: string
          format: uuid
        from_user_id:
          type: string
          format: uuid
        to_user_id:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
    LikeRequest:
      x-go-type: dto.LikeRequest
      x-go-type-import:
        name: dto
        path: github.com/kitoyanok66/workk/dto
      type: object
      description: Request body for creating a like or dislike.
      properties:
        from_user_id:
          type: string
          format: uuid
        to_user_id:
          type: string
          format: uuid
    LikeResponse:
      x-go-type: dto.LikeResponse
      x-go-type-import:
        name: dto
        path: github.com/kitoyanok66/workk/dto
      type: object
      description: Response body for creating a like and next card.
      properties:
        like:
          $ref: '#/components/schemas/LikeDTO'
        match:
          $ref: '#/components/schemas/MatchDTO'
          nullable: true
        next:
          type: object
          description: Next card (freelancer or project)
    DislikeResponse:
      x-go-type: dto.DislikeResponse
      x-go-type-import:
        name: dto
        path: github.com/kitoyanok66/workk/dto
      type: object
      description: Response body for creating a dislike and next card.
      properties:
        next:
          type: object
          description: Next card (freelancer or project)
    ProjectDTO:
      x-go-type: dto.ProjectDTO
      x-go-type-import:
        name: dto
        path: github.com/kitoyanok66/workk/dto
      type: object
      description: Data Transfer Object representing a project.
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        budget:
          type: number
          format: float
        deadline:
          type: string
          format: date-time
        skills:
          type: array
          items:
            $ref: '#/components/schemas/SkillDTO'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    FreelancerDTO:
      x-go-type: dto.FreelancerDTO
      x-go-type-import:
        name: dto
        path: github.com/kitoyanok66/workk/dto
      type: object
      description: Data Transfer Object representing a freelancer.
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        hourly_rate:
          type: number
        portfolio_url:
          type: string
        experience_years:
          type: integer
        rating:
          type: number
        skills:
          type: array
          items:
            $ref: '#/components/schemas/SkillDTO'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    SkillDTO:
      x-go-type: dto.SkillDTO
      x-go-type-import:
        name: dto
        path: github.com/kitoyanok66/workk/dto
      type: object
      description: Data Transfer Object representing a skill.
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        category:
          type: string
        description:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    UserDTO:
      x-go-type: dto.UserDTO
      x-go-type-import:
        name: dto
        path: github.com/kitoyanok66/workk/dto
      type: object
      description: Data Transfer Object representing a user.
      properties:
        id:
          type: string
          format: uuid
        telegram_user_id:
          type: integer
        telegram_username:
          type: string
        full_name:
          type: string
        role:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          format: int
        message:
          type: string